+++ archetypes/plugin.ts
@@ -1,16 +1,20 @@
 import { log } from "../core/utils/log.ts";
 import { toCamelCase, toSnakeCase } from "../deps/text.ts";
+import type { Archetype } from "../core/archetypes.ts";
 
-export default function (name: string) {
+export default (function (name?: string) {
+  name ??= prompt("Name of the plugin:") ?? undefined;
+
   if (!name) {
-    log.error("Missing plugin name. Run 'deno task new plugin plugin-name");
+    log.error("Missing name argument. Run 'deno task new plugin {name}");
     return;
   }
 
   const filename = toSnakeCase(name);
-  const pluginName = toCamelCase(name);
+  const fnName = toCamelCase(name);
 
   return {
+    base: "root",
     path: `/_plugins/${filename}.ts`,
     content: `import { merge } from "lume/core/utils/object.ts";
 
@@ -26,15 +30,14 @@ export interface Options {
 export const defaults = {
 } satisfies Options;
 
-/** ${name} plugin */
-export function ${pluginName}(userOptions?: Options) {
+export function ${fnName}(userOptions?: Options) {
   const options = merge(defaults, userOptions);
 
-  return <D extends PluginData>(site: Site<D>) => {
-  }
-}
-
-export default ${pluginName};
-`,
+  return <D extends PluginData>(site: Lume.Site<D>) => {
   };
 }
+
+export default ${fnName};
+`,
+  };
+}) satisfies Archetype;